The interface is placed in a separate module. The interface has been adjusted.
