+++ app_data/language.txt
@@ -1,1 +1,1 @@
-english
+russian
